+++ server/src/main/java/br/ufal/ic/odontolog/controllers/ReviewableController.java
@@ -1,9 +1,7 @@
 package br.ufal.ic.odontolog.controllers;
 
 import br.ufal.ic.odontolog.api.ReviewableApi;
-import br.ufal.ic.odontolog.dtos.ReviewableCurrentSupervisorFilterDTO;
-import br.ufal.ic.odontolog.dtos.ReviewableDTO;
-import br.ufal.ic.odontolog.dtos.ReviewersDTO;
+import br.ufal.ic.odontolog.dtos.*;
 import br.ufal.ic.odontolog.services.ReviewableService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -58,5 +56,12 @@ public class ReviewableController implements ReviewableApi {
 
     var updated = reviewableService.updateSupervisorsFromReviewables(reviewableId, request);
     return ResponseEntity.ok(updated);
+  }
+
+  @PatchMapping("/{reviewableId}/notes")
+  @PreAuthorize("hasAnyRole('STUDENT','SUPERVISOR')")
+  public ReviewableDTO updateNotes(
+      @PathVariable Long reviewableId, @RequestBody UpdateNotesRequestDTO request) {
+    return reviewableService.updateNotes(reviewableId, request.getNotes());
   }
 }
