+++ server/src/main/java/br/ufal/ic/odontolog/controllers/TreatmentPlanController.java
@@ -6,7 +6,6 @@ import br.ufal.ic.odontolog.dtos.TreatmentPlanAssignUserRequestDTO;
 import br.ufal.ic.odontolog.dtos.TreatmentPlanDTO;
 import br.ufal.ic.odontolog.dtos.TreatmentPlanShortDTO;
 import br.ufal.ic.odontolog.dtos.TreatmentPlanSubmitForReviewDTO;
-import br.ufal.ic.odontolog.dtos.UpdateNotesRequestDTO;
 import br.ufal.ic.odontolog.services.TreatmentPlanService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -61,12 +60,5 @@ public class TreatmentPlanController implements TreatmentPlanApi {
   @PreAuthorize("hasAnyRole('STUDENT','SUPERVISOR')")
   public List<TreatmentPlanShortDTO> getTreatmentPlansByPatient(@PathVariable Long patientId) {
     return treatmentPlanService.getTreatmentPlansByPatientId(patientId);
-  }
-
-  @PatchMapping("/treatment-plan/{id}/notes")
-  @PreAuthorize("hasAnyRole('STUDENT','SUPERVISOR')")
-  public TreatmentPlanDTO updateNotes(
-      @PathVariable Long id, @RequestBody UpdateNotesRequestDTO request) {
-    return treatmentPlanService.updateNotes(id, request.getNotes());
   }
 }
