+++ server/src/test/java/br/ufal/ic/odontolog/controllers/ReviewableControllerIntegrationTest.java
@@ -1,0 +1,127 @@
+package br.ufal.ic.odontolog.controllers;
+
+import br.ufal.ic.odontolog.dtos.ReviewableDTO;
+import br.ufal.ic.odontolog.dtos.TreatmentPlanDTO;
+import br.ufal.ic.odontolog.models.Patient;
+import br.ufal.ic.odontolog.models.Supervisor;
+import br.ufal.ic.odontolog.repositories.PatientRepository;
+import br.ufal.ic.odontolog.repositories.SupervisorRepository;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.transaction.annotation.Transactional;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles("test")
+@Transactional
+public class ReviewableControllerIntegrationTest {
+    @Autowired
+    MockMvc mockMvc;
+    @Autowired
+    PatientRepository patientRepository;
+    @Autowired private SupervisorRepository supervisorRepository;
+    @Autowired
+    ObjectMapper objectMapper;
+
+    private Patient patient;
+    private Supervisor supervisor;
+
+    @BeforeEach
+    void setupPatient() {
+        patient = patientRepository.save(Patient.builder().name("Patient_Test_001").build());
+
+        supervisor =
+                supervisorRepository
+                        .findByEmail("supervisor@test.com")
+                        .orElseGet(
+                                () ->
+                                        supervisorRepository.save(
+                                                Supervisor.builder()
+                                                        .name("supervisor 1")
+                                                        .email("supervisor@test.com")
+                                                        .build()));
+    }
+
+    @Test
+    @WithMockUser(
+            username = "supervisor@test.com",
+            roles = {"SUPERVISOR"})
+    void updateNotes_createsActivityAndUpdatesNotes() throws Exception {
+        var createBody = """
+        {"patientId":"%s"}
+        """.formatted(patient.getId());
+
+        String postJson =
+                mockMvc
+                        .perform(post("/api/treatment-plan").contentType(APPLICATION_JSON).content(createBody))
+                        .andExpect(status().isOk())
+                        .andReturn()
+                        .getResponse()
+                        .getContentAsString();
+        TreatmentPlanDTO created = objectMapper.readValue(postJson, TreatmentPlanDTO.class);
+
+        var patchBody = """
+        {"notes": "New notes value"}
+        """;
+        String patchJson =
+                mockMvc
+                        .perform(
+                                org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch(
+                                                "/api/reviewables/" + created.getId() + "/notes")
+                                        .contentType(APPLICATION_JSON)
+                                        .content(patchBody))
+                        .andExpect(status().isOk())
+                        .andReturn()
+                        .getResponse()
+                        .getContentAsString();
+        ReviewableDTO updated = objectMapper.readValue(patchJson, ReviewableDTO.class);
+        assertThat(updated.getNotes()).isEqualTo("New notes value");
+
+        var activities = updated.getHistory();
+        activities.sort((a, b) -> a.getCreatedAt().compareTo(b.getCreatedAt()));
+        assertThat(activities.size()).isGreaterThan(0);
+        var lastActivity = activities.get(activities.size() - 1);
+        assertThat(lastActivity.getDescription()).isEqualTo("Notes updated by user "+ supervisor.getName());
+
+        var metadata = lastActivity.getMetadata();
+
+        assertThat(metadata.get("data")).isEqualTo("New notes value");
+        assertThat(metadata.get("oldData")).isNull();
+
+        // New round to ensure oldData is also captured
+        patchBody = """
+        {"notes": "New notes value 2"}
+        """;
+        patchJson =
+                mockMvc
+                        .perform(
+                                org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch(
+                                                "/api/reviewables/" + created.getId() + "/notes")
+                                        .contentType(APPLICATION_JSON)
+                                        .content(patchBody))
+                        .andExpect(status().isOk())
+                        .andReturn()
+                        .getResponse()
+                        .getContentAsString();
+        updated = objectMapper.readValue(patchJson, ReviewableDTO.class);
+        assertThat(updated.getNotes()).isEqualTo("New notes value 2");
+
+        activities = updated.getHistory();
+        activities.sort((a, b) -> a.getCreatedAt().compareTo(b.getCreatedAt()));
+        assertThat(activities.size()).isGreaterThan(1);
+        lastActivity = activities.get(activities.size() - 1);
+        assertThat(lastActivity.getDescription()).isEqualTo("Notes updated by user " + supervisor.getName());
+    }
+}
