+++ server/src/test/java/br/ufal/ic/odontolog/controllers/TreatmentPlanControllerIntegrationTest.java
@@ -219,75 +219,4 @@ class TreatmentPlanControllerIntegrationTest {
     assertThat(dtos.stream().allMatch(dto -> dto.getPatient().getId().equals(patient.getId())))
         .isTrue();
   }
-
-  @Test
-  @WithMockUser(
-      username = "supervisor@test.com",
-      roles = {"SUPERVISOR"})
-  void updateNotes_createsActivityAndUpdatesNotes() throws Exception {
-    var createBody = """
-        {"patientId":"%s"}
-        """.formatted(patient.getId());
-
-    String postJson =
-        mockMvc
-            .perform(post("/api/treatment-plan").contentType(APPLICATION_JSON).content(createBody))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-    TreatmentPlanDTO created = objectMapper.readValue(postJson, TreatmentPlanDTO.class);
-
-    var patchBody = """
-        {"notes": "New notes value"}
-        """;
-    String patchJson =
-        mockMvc
-            .perform(
-                org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch(
-                        "/api/treatment-plan/" + created.getId() + "/notes")
-                    .contentType(APPLICATION_JSON)
-                    .content(patchBody))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-    TreatmentPlanDTO updated = objectMapper.readValue(patchJson, TreatmentPlanDTO.class);
-    assertThat(updated.getNotes()).isEqualTo("New notes value");
-
-    var activities = updated.getHistory();
-    activities.sort((a, b) -> a.getCreatedAt().compareTo(b.getCreatedAt()));
-    assertThat(activities.size()).isGreaterThan(0);
-    var lastActivity = activities.get(activities.size() - 1);
-    assertThat(lastActivity.getDescription()).isEqualTo("Notes updated by user supervisor 1");
-
-    var metadata = lastActivity.getMetadata();
-
-    assertThat(metadata.get("data")).isEqualTo("New notes value");
-    assertThat(metadata.get("oldData")).isNull();
-
-    // New round to ensure oldData is also captured
-    patchBody = """
-        {"notes": "New notes value 2"}
-        """;
-    patchJson =
-        mockMvc
-            .perform(
-                org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch(
-                        "/api/treatment-plan/" + created.getId() + "/notes")
-                    .contentType(APPLICATION_JSON)
-                    .content(patchBody))
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-    updated = objectMapper.readValue(patchJson, TreatmentPlanDTO.class);
-    assertThat(updated.getNotes()).isEqualTo("New notes value 2");
-
-    activities = updated.getHistory();
-    activities.sort((a, b) -> a.getCreatedAt().compareTo(b.getCreatedAt()));
-    assertThat(activities.size()).isGreaterThan(1);
-    lastActivity = activities.get(activities.size() - 1);
-    assertThat(lastActivity.getDescription()).isEqualTo("Notes updated by user supervisor 1");
-  }
 }
